Split CanvasPage apart
