+++ blueprint/src/app/canvas/page.js
@@ -2,13 +2,13 @@
 
 import { useEffect, useState, useRef } from "react";
 import styles from './page.module.css';
-import { WidgetRenderer } from './WidgetRenderer';
-import { RenderPage } from "./HtmlConverter";
 import {
-  TransformWrapper,
-  TransformComponent,
   useControls,
 } from "react-zoom-pan-pinch";
+
+import { Canvas } from './Canvas';
+import { LeftPanel } from './LeftPanel';
+import { RightPanel } from './RightPanel';
 
 export default function CanvasPage() {
   // Pages, each containing widgets
@@ -315,67 +315,22 @@ export default function CanvasPage() {
         <LeftPanel createWidget={createWidget} />
       </div>
 
-      <TransformWrapper
-        initialScale={1}
-        initialPositionX={0}
-        initialPositionY={0}
-        disabled={isPlacing || isDragging}
-        limitToBounds={false}
-        panning={{ velocityDisabled: true }}
-        minScale={0.05}
-        // Keep track of zoom transform and scale
-        onTransformed={({ state }) => {
-          setScale(state.scale);
-          setTransformCoords({posX: state.positionX, posY: state.positionY});
-        }}
-      >
-        {/* Canvas area, a window to view the current page */}
-        <div className={styles.canvasArea}>
-
-          <TransformComponent>
-
-            {/* The component for moving/zooming the camera */}
-            <div className={styles.canvasView} ref={canvasRef} onClick={handleCanvasClick}>
-                
-                {/* The page itself is its own component here */}
-                <div className={styles.pages}
-                  style={{
-                    width: currentPage?.width + "px",
-                    height: currentPage?.height + "px",
-                    backgroundColor: currentPage.backgroundColor,
-                  }}
-                >
-                  {/* Render new objects, only if widgets exists */}
-                  {Array.isArray(widgets) &&
-                    widgets.map((widget) => (
-                      <WidgetRenderer
-                        bounds="parent"
-                        key={widget.id}
-                        widget={widget}
-                        // selectedWidgets? means if selectedWidgets is not null
-                        // .some checks if any widgets in the array have the same id
-                        isSelected={selectedWidgets?.some((w) => w.id === widget.id)}
-                        onClick={() => {
-                          // Select the widget when clicked
-                          setSelectedWidgets([widget]);
-                          console.log("Selected widget: " + widget.id);
-                        }}
-                        onDragStart={() => setIsDragging(true)}
-                        onDragStop={() => setIsDragging(false)}
-                        alertDragStop={updateWidget}
-                        scale={scale}
-                      />
-                    ))}
-
-                  {/* If placing a widget, render it at the mouse position */}
-                  {isPlacing && widgetToPlace && (
-                    <WidgetRenderer key={"placing-" + widgetToPlace.id} widget={widgetToPlace} scale={scale} />
-                  )}
-                </div>
-            </div>
-            </TransformComponent>
-        </div>
-      </TransformWrapper>
+      <Canvas
+        widgets={widgets}
+        isPlacing={isPlacing}
+        isDragging={isDragging}
+        widgetToPlace={widgetToPlace}
+        selectedWidgets={selectedWidgets}
+        setSelectedWidgets={setSelectedWidgets}
+        setIsDragging={setIsDragging}
+        updateWidget={updateWidget}
+        scale={scale}
+        setScale={setScale}
+        setTransformCoords={setTransformCoords}
+        currentPage={currentPage}
+        canvasRef={canvasRef}
+        handleCanvasClick={handleCanvasClick}
+      />
 
       {/* Panel on the right, showing options for the selected widgets and the canvas page */}
       <div className={`${styles.rightPanel} ${styles.sidePanel}`}>
@@ -404,398 +359,4 @@ export default function CanvasPage() {
 
     setPages(changedPages);
   }
-}
-
-function LeftPanel({ createWidget }) {
-    return (
-      <div>
-          <div className={styles.sectionTitle}>Objects</div>
-          <button className={styles.categoryItem}>Text Box</button>
-          <button className={styles.categoryItem}>Image</button>
-
-          <div className={styles.divider}></div>
-
-          <div className={styles.sectionTitle}>Widgets</div>
-          <button className={styles.categoryItem} onClick={() => createWidget('video')}>Video</button>
-          <button className={styles.categoryItem} onClick={() => createWidget('dropdown')}>Dropdown</button>
-          <button className={styles.categoryItem} onClick={() => createWidget('advert')}>Advertisement</button>
-
-          <div className={styles.divider}></div>
-
-          <div className={styles.sectionTitle}>Shapes</div>
-          <button className={styles.categoryItem} onClick={() => createWidget('box')}>Box</button>
-          <button className={styles.categoryItem}>Forms</button>
-          <button className={styles.categoryItem}>Images</button>
-      </div>
-    );
-  }
-
-  function RightPanel({
-    changeWidgetProperty, selectedWidgets, widgets, deleteWidget,
-    pages, selectedPageID, setSelectedPageID, currentPage, createPage, changePageProperty
-   }) {
-    const [buttonSelected, setButtonSelected] = useState(false);
-
-    const handleButtonClick = () => {
-      setButtonSelected(!buttonSelected);
-      console.log(`Canvas status: ${!buttonSelected}`);
-    }
-
-    // Return a button to switch between the selected widgets and canvas settings
-    return (
-      <div>
-        {/* Button to switch, always rendered. When clicked, invert buttonSelected */}
-        <button className={styles.switchButton} onClick={handleButtonClick}>Switch</button>
-
-        {/* False = widget properties. True = page properties */}
-        {!buttonSelected ? <RightWidgetPanel changeWidgetProperty={changeWidgetProperty} selectedWidgets={selectedWidgets} widgets={widgets} deleteWidget={deleteWidget} />
-          : <RightPagePanel pages={pages} selectedPageID={selectedPageID} setSelectedPageID={setSelectedPageID} currentPage={currentPage} createPage={createPage} changePageProperty={changePageProperty} />}
-      </div>
-
-    );
-  }
-
-  function RightWidgetPanel({ changeWidgetProperty, selectedWidgets, widgets, deleteWidget }) {
-    // Render the selected widgets panel
-    if (selectedWidgets && selectedWidgets.length > 0) {
-    // If something is selected
-      return (
-        <div>
-          <div>
-            {/* Button for deleting widgets */}
-            <button
-              className={styles.deleteButton}
-              onClick={() => {
-                selectedWidgets.forEach(widget => {
-                  console.log('Deleting ', widget)
-                  deleteWidget(widget.id);
-                })
-              }}
-            >Delete Selected Widget</button>
-          </div>
-  
-          {/* Menu to change widgets */}
-          {selectedWidgets.map((widget) => (
-            <div key={widget.id} className={styles.widgetOptions}>
-              {/* Inputs to change widget properties */}
-              <p>Color:</p>
-              <input
-                type="color"
-                value={widget.backgroundColor || "#cccccc"}
-                onChange={e => changeWidgetProperty(widget.id, { backgroundColor: e.target.value })}
-              />
-  
-              {/* Two rotation inputs, one as an editable display and the other as a slider */}
-              <p>Rotation: {widget.rotation}</p>
-              <input
-                type="number"
-                min="0"
-                max="360"
-                value={widget.rotation || 0}
-                onChange={e =>
-                  changeWidgetProperty(widget.id, { rotation: parseInt(e.target.value, 10) })
-                }/>
-              <input
-                type="range"
-                min="0"
-                max="360"
-                value={widget.rotation || 0}
-                onChange={e =>
-                  changeWidgetProperty(widget.id, { rotation: parseInt(e.target.value || "0", 10) })
-                }
-              />
-  
-              {/* ===== Video controls ===== */}
-              {widget.type === 'video' && (
-                <>
-                  <p>Video URL:</p>
-                  <input
-                    type="text"
-                    placeholder="/videos/intro.mp4 or https://cdn.example.com/clip.mp4"
-                    value={widget.videoUrl || ""}
-                    onChange={e =>
-                      changeWidgetProperty(widget.id, { videoUrl: e.target.value || null })
-                    }
-                  />
-  
-                  <label
-                  style={{
-                    display: "flex",
-                    alignItems: "center",
-                    justifyContent: "space-between",
-                    gap: 8,
-                  }}
-                >
-                  <span>Loop</span>
-                  <input
-                    type="checkbox"
-                    checked={!!widget.loop}
-                    onChange={e => changeWidgetProperty(widget.id, { loop: e.target.checked })}
-                  />
-                </label>
-
-                <label
-                  style={{
-                    display: "flex",
-                    alignItems: "center",
-                    justifyContent: "space-between",
-                    gap: 8,
-                  }}
-                >
-                  <span>Muted</span>
-                  <input
-                    type="checkbox"
-                    checked={!!widget.muted}
-                    onChange={e => changeWidgetProperty(widget.id, { muted: e.target.checked })}
-                  />
-                </label>
-
-                <label
-                  style={{
-                    display: "flex",
-                    alignItems: "center",
-                    justifyContent: "space-between",
-                    gap: 8,
-                  }}
-                >
-                  <span>Show controls</span>
-                  <input
-                    type="checkbox"
-                    checked={!!widget.controls}
-                    onChange={(e) =>
-                      changeWidgetProperty(widget.id, { controls: e.target.checked })
-                    }
-                  />
-                </label>
-
-                <label
-                  style={{
-                    display: "flex",
-                    alignItems: "center",
-                    justifyContent: "space-between",
-                    gap: 8,
-                  }}
-                >
-                  <span>Autoplay</span>
-                  <input
-                    type="checkbox"
-                    checked={!!widget.autoplay}
-                    onChange={(e) =>
-                      changeWidgetProperty(widget.id, { autoplay: e.target.checked })
-                    }
-                  />
-                </label>
-                </>
-              )}
-  
-              {/* ===== Dropdown controls ===== */}
-              {widget.type === 'dropdown' && (() => {
-                // Use a draft string while typing; default to current options joined with ", "
-                const draft = widget.optionsText ?? (widget.options || []).join(", ");
-
-                // Commit helper: turn the draft string into an options array and normalize it
-                const commitOptions = (raw) => {
-                  const arr = (raw || "")
-                    .split(",")
-                    .map(s => s.trim())
-                    .filter(Boolean); // drop empty tokens on commit only
-
-                  const nextValue = arr.includes(widget.value) ? widget.value : (arr[0] || "");
-                  // Save normalized text too so the input shows tidy commas + spaces
-                  const normalizedText = arr.join(", ");
-
-                  changeWidgetProperty(widget.id, {
-                    options: arr,
-                    value: nextValue,
-                    optionsText: normalizedText,
-                  });
-                };
-
-                return (
-                  <>
-                    <p>Options (comma-separated):</p>
-                    <input
-                      type="text"
-                      value={draft}
-                      onChange={e => {
-                        // While typing, only update the draft string; don't parse/sanitize yet.
-                        changeWidgetProperty(widget.id, { optionsText: e.target.value });
-                      }}
-                      onKeyDown={e => {
-                        if (e.key === "Enter") {
-                          e.preventDefault();
-                          commitOptions(e.currentTarget.value);
-                          // keep focus if you want: e.currentTarget.blur();
-                        }
-                      }}
-                      onBlur={e => commitOptions(e.currentTarget.value)}
-                    />
-
-                    <p>Selected Value:</p>
-                    <input
-                      type="text"
-                      value={widget.value || ""}
-                      onChange={e => changeWidgetProperty(widget.id, { value: e.target.value })}
-                    />
-
-                    <p>Font Size:</p>
-                    <input
-                      type="number"
-                      min="10"
-                      max="48"
-                      value={widget.fontSize || 14}
-                      onChange={e =>
-                        changeWidgetProperty(
-                          widget.id,
-                          { fontSize: parseInt(e.target.value || "14", 10) }
-                        )
-                      }
-                    />
-
-                    <p>Text Color:</p>
-                    <input
-                      type="color"
-                      value={widget.textColor || "#111111"}
-                      onChange={e => changeWidgetProperty(widget.id, { textColor: e.target.value })}
-                    />
-                  </>
-                );
-              })()}
-
-  
-              {/* ===== Advertisement controls ===== */}
-              {widget.type === 'advert' && (
-                <>
-                  <p>Image URL:</p>
-                  <input
-                    type="text"
-                    placeholder="https://…/banner.jpg"
-                    value={widget.imageUrl || ""}
-                    onChange={e => changeWidgetProperty(widget.id, { imageUrl: e.target.value })}
-                  />
-  
-                  <p>Link URL:</p>
-                  <input
-                    type="text"
-                    placeholder="www.google.com"
-                    value={widget.linkUrl || ""}
-                    onChange={e => changeWidgetProperty(widget.id, { linkUrl: e.target.value })}
-                  />
-  
-                  <p>Object Fit:</p>
-                  <select
-                    value={widget.objectFit || "cover"}
-                    onChange={e => changeWidgetProperty(widget.id, { objectFit: e.target.value })}
-                  >
-                    <option value="contain">contain</option>
-                    <option value="cover">cover</option>
-                    <option value="fill">fill</option>
-                    <option value="none">none</option>
-                    <option value="scale-down">scale-down</option>
-                  </select>
-  
-                  <label>
-                    <input
-                      type="checkbox"
-                      checked={!!widget.showBorder}
-                      onChange={e => changeWidgetProperty(widget.id, { showBorder: e.target.checked })}
-                    />
-                    Show border
-                  </label>
-  
-                  <p>Border Color:</p>
-                  <input
-                    type="color"
-                    value={widget.borderColor || "#333333"}
-                    onChange={e => changeWidgetProperty(widget.id, { borderColor: e.target.value })}
-                  />
-                </>
-              )}
-            </div>
-          ))}
-        </div>
-      );
-    }
-  else {
-      // If something is not selected
-      return (
-      <p>Select a widget to view its properties.</p>
-    );
-    }
-  }  
-
-function RightPagePanel({ pages, selectedPageID, setSelectedPageID, currentPage, createPage, changePageProperty }) {
-
-  const downloadHTMLPage = (currentPage) => {
-    console.log("Downloading page", currentPage);
-
-    const file = new Blob([RenderPage(currentPage)], {type: 'text/html'});
-
-    const element = document.createElement("a");
-    element.href = URL.createObjectURL(file);
-    element.download = "Page" + Date.now() + ".html";
-
-    document.body.appendChild(element);
-  }
-
-  return(
-    <div className={styles.widgetOptions}>
-      {/* A view of all pages */}
-      <select value={selectedPageID} onChange={e => setSelectedPageID(Number(e.target.value))}>
-        {pages.map(page => (
-          <option key={page.id} value={page.id}>{page.name}</option>
-        ))}
-      </select>
-      
-      {/* Buttons for editing pages */}
-      <button onClick={createPage}>+ New Page</button>
-
-      {/* Page size controls */}
-      <div>
-        <p>Width</p>
-        <input 
-          type="input"
-          min="0"
-          autoComplete="off"
-          value={currentPage.width}
-          onChange={e => {
-            if (e?.target.value === "") {
-              changePageProperty(selectedPageID, { width: 0 })
-            } else {
-              changePageProperty(selectedPageID, { width: parseInt(e.target.value, 10) })
-            }
-          }}/>
-
-        <p>Height</p>
-        <input 
-          type="input"
-          min="0"
-          autoComplete="off"
-          value={currentPage.height}
-          onChange={e => {
-            if (e?.target.value === "") {
-              changePageProperty(selectedPageID, { height: 0 })
-            } else {
-              changePageProperty(selectedPageID, { height: parseInt(e.target.value, 10) })
-            }
-          }}
-          />
-
-          <p>Background Color</p>
-          <input 
-            type="color"
-            onChange={e => {
-              if (e?.target.value === "") {
-                changePageProperty(selectedPageID, { backgroundColor: '#ffffff' })
-              } else {
-                changePageProperty(selectedPageID, { backgroundColor: e.target.value })
-              }
-            }}
-          />
-      </div>
-
-      {/* Download current page as HTML */}
-      <button >Download Page</button>
-    </div>
-  );
 }
